requirements.txt
fastapi==0.115.0
uvicorn==0.34.0
httpx==0.28.1
pydantic==2.12.0
python-dotenv==1.0.0
openai==1.82.0